% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{calc_mc_pval}
\alias{calc_mc_pval}
\title{Calculate a monte carlo p-value estimation}
\usage{
calc_mc_pval(zscore.df)
}
\arguments{
\item{diff}{A dataframe containing Metabs as rows and conditions as columns}
}
\value{
A dataframe containing estimated p-values for each predicted value
}
\description{
This function uses a z-score prediction table to estimate p-values using monte carlo estimation. It calculates
the significance of a z-score compared to the metabolite's "normal behaviour" over multiple conditions.
}
