% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iem_names}
\alias{iem_names}
\title{IEM names and IDs.}
\format{
A data frame with 49 rows and 4 variables:
\describe{
\item{IEM.full}{The full IEM names, may contain lower cases}
\item{IEM.code}{IEM codes used in sampling for identification}
\item{IEM.full.upper}{The full IEM names with consistent capitalisation}
\item{IEM.short}{Short versions of the IEM names for figures}
}
}
\usage{
iem_names
}
\description{
A dataset containing IEM names and IDs from Fig3 of the Recon 2 article.
These are also stored in the correct order to reproduce Fig 3 from the Recon 2 article.
}
\keyword{datasets}
