% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_sampling_results}
\alias{load_sampling_results}
\title{Read in KO and WT sampling data}
\usage{
load_sampling_results(indir, disease.code, single.wt = F, wt.file = NA)
}
\arguments{
\item{indir}{Path to the input dir, should contain a ko folder and wt folder if not using single.wt}

\item{disease.code}{Three-letter code designating the requested disease KO to load}

\item{single.wt}{If True, uses a non-disease-specific WT (provide the file using wt.file)}

\item{wt.file}{Use this to provide the wt path and filename when using single.wt=T}
}
\value{
A list containing ko and wt dataframes of all samples
}
\description{
This function reads in KO and WT sampling data. It assumes that the results files are prefixed
using a 3 letter code (xaa, xab..) and that a code-to-disease dictionary is provided. It supports
using a single WT file or a specific WT file. Files should be in a /ko/ and /wt/ directories.
}
