% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metab_names}
\alias{metab_names}
\title{Metabolite names and IDs.}
\format{
A data frame with 54 rows and 3 variables:
\describe{
\item{Metab.names}{The metabolite names}
\item{Metab.Recon1.IDs}{Metabolite IDs from Recon 1}
\item{Metab.Recon2v2.IDs}{Metabolite IDs from Recon 2.2}
\item{Metab.EX.IDs}{Metabolite IDs in the EX_xxx(e) format}
}
}
\usage{
metab_names
}
\description{
A dataset containing metabolite names and IDs from Recon 1 and 2.2.
These are also stored in the correct order to reproduce Fig 3 from the Recon 2 article.
}
\keyword{datasets}
