% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{obs}
\alias{obs}
\title{Observed metabolite changes in IEM patients.}
\format{
A data frame with 54 rows and 50 variables:
\describe{
\item{Metab}{All 54 metabolites from Fig 3 of the Recon 2 article}
\item{IEM names}{All the other columns are the 49 IEMs from Fig 3 of the Recon 2 article}
}
}
\usage{
obs
}
\description{
A dataset containing the observed metabolite changes in patients from Fig3 of the Recon 2 article.
}
\keyword{datasets}
