% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_precision_recall_curve}
\alias{plot_precision_recall_curve}
\title{Plot a precision-recall curve for results outputted by stat_pr_curve}
\usage{
plot_precision_recall_curve(
  pr.re,
  title = "Default title",
  method = c("merge_pos", "separate_dir", "only_pos")
)
}
\arguments{
\item{pr.re}{A data frame with a column for precision and a column for recall}
}
\description{
Plot a precision-recall curve for results outputted by stat_pr_curve
}
