% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_roc_curve}
\alias{plot_roc_curve}
\title{Plot a ROC curve for results outputted by stat_roc_curve}
\usage{
plot_roc_curve(
  roc,
  title = "Default title",
  method = c("merge_pos", "separate_dir", "only_pos")
)
}
\arguments{
\item{roc}{A data frame with a column for FPR and a column for TPR}
}
\description{
Plot a ROC curve for results outputted by stat_roc_curve
}
