% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{stat_counts}
\alias{stat_counts}
\title{Calculate precision and recall for predicted and observed results}
\usage{
stat_counts(pred.num, obs)
}
\arguments{
\item{pred.num}{A data frame containing numeric prediction values with metabolites as rows and conditions as columns}

\item{obs}{An data frame with metabolites as rows and conditions as columns}
}
\value{
A named vector containing precision and recall
}
\description{
Calculate precision and recall for predicted and observed results
}
